requirements.txt
accelerate==1.10.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
airportsdata==20250909
annotated-types==0.7.0
anthropic==0.37.1
anyio==4.11.0
astor==0.8.1
async-timeout==5.0.1
attrs==25.3.0
blake3==1.0.7
cachetools==6.2.0
cbor2==5.7.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
cloudpickle==3.1.1
cohere==5.14.0
coloredlogs==15.0.1
compressed-tensors==0.9.2
cupy-cuda12x==13.6.0
dacite==1.8.1
dataclasses-json==0.6.7
datasets==3.2.0
deepspeed==0.18.1
depyf==0.18.0
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
dnspython==2.8.0
einops==0.8.1
email-validator==2.3.0
eval_type_backport==0.2.2
fastapi==0.118.0
fastapi-cli==0.0.13
fastapi-cloud-cli==0.2.1
fastavro==1.12.0
fastrlock==0.8.3
filelock==3.19.1
flatbuffers==25.9.23
frozendict==2.4.6
frozenlist==1.7.0
fschat==0.2.36
fsspec==2024.9.0
ftfy==6.3.1
gguf==0.10.0
gitdb==4.0.12
GitPython==3.1.45
h11==0.16.0
hf-xet==1.1.10
hjson==3.1.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.35.3
humanfriendly==10.0
idna==3.10
interegular==0.3.3
Jinja2==3.1.6
jiter==0.11.0
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
lark==1.2.2
latex2mathml==3.78.1
litellm==1.56.2
llama_cpp_python==0.2.76
llguidance==0.7.30
llm-blender==0.0.2
llmlingua==0.2.2
llvmlite==0.43.0
lm-format-enforcer==0.10.12
markdown-it-py==4.0.0
markdown2==2.5.4
MarkupSafe==3.0.3
marshmallow==3.26.1
mdurl==0.1.2
mistral_common==1.8.5
mpmath==1.3.0
msgpack==1.1.1
msgspec==0.19.0
multidict==6.6.4
multiprocess==0.70.16
mypy_extensions==1.1.0
networkx==3.4.2
nh3==0.3.0
ninja==1.13.0
nltk==3.9.2
numba==0.60.0
numpy==1.26.4
# nvidia-cublas-cu11==11.11.3.6
# nvidia-cublas-cu12==12.6.4.1
# nvidia-cuda-cupti-cu11==11.8.87
# nvidia-cuda-cupti-cu12==12.6.80
# nvidia-cuda-nvrtc-cu11==11.8.89
# nvidia-cuda-nvrtc-cu12==12.6.77
# nvidia-cuda-runtime-cu11==11.8.89
# nvidia-cuda-runtime-cu12==12.6.77
# nvidia-cudnn-cu11==9.1.0.70
# nvidia-cudnn-cu12==9.5.1.17
# nvidia-cufft-cu11==10.9.0.58
# nvidia-cufft-cu12==11.3.0.4
# nvidia-cufile-cu12==1.11.1.6
# nvidia-curand-cu11==10.3.0.86
# nvidia-curand-cu12==10.3.7.77
# nvidia-cusolver-cu11==11.4.1.48
# nvidia-cusolver-cu12==11.7.1.2
# nvidia-cusparse-cu11==11.7.5.86
# nvidia-cusparse-cu12==12.5.4.2
# nvidia-cusparselt-cu12==0.6.3
# nvidia-ml-py==13.580.82
# nvidia-nccl-cu11==2.21.5
# nvidia-nccl-cu12==2.26.2
# nvidia-nvjitlink-cu12==12.6.85
# nvidia-nvtx-cu11==11.8.86
# nvidia-nvtx-cu12==12.6.77
onnxruntime==1.19.2
openai==1.56.0
openai-harmony==0.0.4
opencv-python-headless==4.11.0.86
outlines==0.1.11
outlines_core==0.1.26
pandas==2.2.3
partial-json-parser==0.2.1.1.post6
peft==0.17.1
pillow==11.1.0
prettytable==3.11.0
prometheus-fastapi-instrumentator==7.1.0
prometheus_client==0.23.1
propcache==0.3.2
protobuf==6.32.1
py-cpuinfo==9.0.0
pyarrow==21.0.0
pybase64==1.4.2
pycountry==24.6.1
pycparser==2.23
pydantic==2.11.9
pydantic-extra-types==2.10.5
pydantic_core==2.33.2
python-dotenv==1.1.1
python-json-logger==3.3.0
python-multipart==0.0.20
# pytorch-triton-rocm==3.2.0
pytz==2025.2
PyYAML==6.0.3
rank-bm25==0.2.2
ray==2.49.2
referencing==0.36.2
regex==2025.9.18
requests==2.32.4
rich==13.9.4
rich-toolkit==0.15.1
rignore==0.6.4
rpds-py==0.27.1
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.15.3
sentence-transformers==3.3.0
sentencepiece==0.2.0
sentry-sdk==2.39.0
setproctitle==1.3.7
shellingham==1.5.4
shortuuid==1.0.13
smmap==5.0.2
sniffio==1.3.1
soundfile==0.13.1
soxr==1.0.0
starlette==0.48.0
svgwrite==1.4.3
sympy==1.13.1
tabulate==0.9.0
threadpoolctl==3.6.0
tiktoken==0.11.0
together==1.5.3
tokenizers==0.22.1
# torch==2.8.0+cu118
# torchaudio==2.8.0+cu118
torchdata==0.11.0
# torchvision==0.21.0+cu118
tqdm==4.66.5
transformers==4.56.2
triton==3.2.0
typer==0.15.1
types-requests==2.32.4.20250913
typing-inspect==0.9.0
typing-inspection==0.4.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.37.0
uvloop==0.21.0

wandb==0.22.2
watchfiles==1.1.0
wavedrom==2.0.3.post3
websockets==15.0.1
xformers==0.0.29.post2
xgrammar==0.1.21
xxhash==3.6.0
yarl==1.20.1